Testing directory moved. some minor refactoring.
