+++ hw-project/src/test/resources/empty_tracker.config
